move classes to new packages, add release info
