+++ src/extension/geopkg-output/src/test/java/org/geoserver/geopkg/GeoPackageGetFeatureOutputFormatTest.java
@@ -15,8 +15,10 @@ import java.io.IOException;
 import net.opengis.wfs.GetFeatureType;
 import net.opengis.wfs.WfsFactory;
 import org.geoserver.data.test.SystemTestData;
+import org.geoserver.geopackage.GeoPkg;
 import org.geoserver.platform.Operation;
 import org.geoserver.wfs.WFSTestSupport;
+import org.geoserver.wfs.geopackage.GeoPackageGetFeatureOutputFormat;
 import org.geoserver.wfs.request.FeatureCollectionResponse;
 import org.geotools.data.FeatureSource;
 import org.geotools.data.memory.MemoryFeatureCollection;
