+++ src/extension/geopkg-output/src/test/java/org/geoserver/geopkg/GeoPkgDataStoreFactoryInitializerTest.java
@@ -10,6 +10,7 @@ import static org.junit.Assert.assertNotNull;
 
 import java.io.File;
 import org.geoserver.data.DataStoreFactoryInitializer;
+import org.geoserver.geopackage.GeoPkgDataStoreFactoryInitializer;
 import org.geoserver.platform.GeoServerExtensions;
 import org.geoserver.platform.GeoServerResourceLoader;
 import org.geotools.geopkg.GeoPkgDataStoreFactory;
