+++ src/extension/pom.xml
@@ -404,6 +404,13 @@
         <module>mapml</module>
       </modules>
     </profile>
+    <profile>
+      <id>geopkg-output</id>
+      <modules>
+        <module>geopkg-output</module>
+      </modules>
+    </profile>
+
     <!-- allExtensions profile, includes all modules -->
     <profile>
       <id>allExtensions</id>
@@ -452,6 +459,7 @@
         <module>gwc-s3</module>
         <module>wps-download</module>
         <module>mapml</module>
+        <module>geopkg-output</module>
       </modules>
     </profile>
