+++ src/release/ext-geopkg-output.xml
@@ -1,0 +1,16 @@
+<assembly>
+    <id>geopkg-output-plugin</id>
+    <formats>
+        <format>zip</format>
+    </formats>
+    <includeBaseDirectory>false</includeBaseDirectory>
+    <fileSets>
+        <fileSet>
+            <directory>release/target/dependency</directory>
+            <outputDirectory></outputDirectory>
+            <includes>
+                <include>gs-geopkg-output*.jar</include>
+            </includes>
+        </fileSet>
+    </fileSets>
+</assembly>
