+++ src/release/pom.xml
@@ -153,6 +153,11 @@
     </dependency>
     <dependency>
       <groupId>org.geoserver.extension</groupId>
+      <artifactId>gs-geopkg-output</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.geoserver.extension</groupId>
       <artifactId>gs-ogr-wfs</artifactId>
       <version>${project.version}</version>
     </dependency>
